trusted-types/block-string-assignment-to-Location-replace.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>

  <meta http-equiv="Content-Security-Policy" content="trusted-types">
</head>
<body>
<script>
  // TrustedURL replacements do not throw.
  test(t => {
    let p = createURL_policy(window, 1);
    let url = p.createURL(location.href + "#xxx");
    location.replace(url);
    assert_equals("" + url, location.href, "location href");
  }, "location.replace via policy (successful URL transformation).");

  // String replacements throw.
  test(t => {
    assert_throws(new TypeError(), _ => {
      location.replace("A string");
    });
  }, "`location.replace = string` throws");

  // Null replacement throws.
  test(t => {
    assert_throws(new TypeError(), _ => {
      location.replace(null);
    });
  }, "`location.replace = null` throws");

  // After default policy creation string assignment implicitly calls createURL.
  test(t => {
    let p = window.TrustedTypes.createPolicy("default", { createURL: createURLJS }, true);
    location.replace(INPUTS.URL);
    assert_equals(RESULTS.URL, location.href, "location href");
  }, "`location.replace = string` via default policy (successful URL transformation).");

  // After default policy creation null assignment implicitly calls createURL.
  test(t => {
    location.replace(null);
    assert_equals("null", location.href, "location href");
  }, "`location.replace = null` via default policy (successful URL transformation).");
</script>
